refactor(renderer): give conversation its types beside its logic, not in the view
`ConversationView.tsx` declared `ConversationProps`, `ComposerDraft`,
`ConversationTarget`, `MediaPreview`, `SidebarFilePreview` and `RightPanelMode`,
and fifteen modules in the feature imported them from there: eleven stores, the
scope, the controller, the draft logic and the timeline. None of them renders
anything, so each was reaching through a component to borrow a name.

It also closed a loop. `ConversationView.tsx` imports `conversation-scope.ts`
for its value, and `conversation-scope.ts` imported `ConversationView` back for
its types. That is a cycle `noImportCycles` never reported, because one
direction is `import type` and it only follows value edges.

`conversation-types.ts` holds the six declarations and imports nothing from the
feature, which breaks the loop rather than hiding it. This is the same shape as
`sidebar-types.ts` and the last outstanding item of the migration plan.

Pure move: no declaration changed, no test assertion changed. Renderer only.
`typecheck:renderer`, `check:ui` and the 207 conversation tests are green.
