+++ src/renderer/src/features/conversation/ConversationTimeline.tsx
@@ -14,8 +14,8 @@ import { AttachmentCards } from "./AttachmentCards";
 import { ChatActionMarker } from "./ChatActionMarker";
 import { ChatSearch } from "./ChatSearch";
 import { BrowserTakeoverCard } from "./ConversationPrompts";
-import type { ConversationProps } from "./ConversationView";
 import { useConversationViewScope } from "./conversation-scope";
+import type { ConversationProps } from "./conversation-types";
 import { messageContentBlocks } from "./DataTable";
 import { ScrollToLatestButton } from "./MessageNavigation";
 import { MessageActions, MessageBody } from "./MessageRendering";
